# docs/openapi.yaml
openapi: 3.0.3
info:
  title: G2M Backend API
  version: 1.0.0
  description: |
    REST API for the G2M e‑commerce platform. Includes authentication, user profile management,
    products, orders, reviews, vouchers, and admin endpoints. JWT-based auth with refresh sessions,
    Cloudinary uploads, and Nodemailer emails.
servers:
  - url: http://localhost:5000/api
    description: Local
  - url: https://your-deployment-domain/api
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        _id: { type: string }
        sku: { type: string }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        subcategory: { type: string }
        price: { type: number }
        discountPrice: { type: number, nullable: true }
        mainImage: { type: string }
        images: { type: array, items: { type: string } }
        avgRating: { type: number }
        numReviews: { type: number }
...
security:
  - bearerAuth: []
paths:
...
  /products/{productId}:
    get:
      tags: [Products]
      summary: Get product by id
      security: []
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
...
  /orders/{orderNumber}:
    get:
      tags: [Orders]
      summary: Get an order by number
      parameters:
        - in: path
          name: orderNumber
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    patch:
      tags: [Orders]
      summary: Cancel an order
      parameters:
        - in: path
          name: orderNumber
          required: true
          schema: { type: string }
      responses:
        '200': { description: Cancelled }
...